Partial rewrite to use the new interchange format everywhere.
