+++ primap2/_data_format.py
@@ -133,7 +133,11 @@ class DatasetDataFormatAccessor(_accessor_base.BaseDatasetAccessor):
 
         df = pd.concat(dfs, ignore_index=True)
 
-        df = pm2io._data_reading.sort_columns_and_rows(df)
+        df = pm2io._data_reading.sort_columns_and_rows(
+            df,
+            dimensions=[dim for dim in dsd.dims if dim != "time"]
+            + [entity_col, "unit"],
+        )
 
         df.attrs = {
             "attrs": self._ds.attrs,
