--- primap2/tests/data/test_convert_dataframe_units_primap_to_primap2.csv
@@ -1,12 +1,0 @@
-,country,category,classification,gas,unit,1991,2000,2010
-0,AUS,IPC1,TOTAL,CO2,Mt CO2 / yr,4,5,6
-1,AUS,IPC0,TOTAL,KYOTOGHG,Mt CO2 / yr,8,9,10
-2,ZAM,IPC0,TOTAL,KYOTOGHG,Gg CO2 / yr,30,20,40
-3,ZAM,IPC2,TOTAL,CH4,Gg CH4 / yr,7,8,9
-4,ZAM,IPC2,TOTAL,CO2,Gg CO2 / yr,12,13,14
-5,USA,IPC0,TOTAL,KYOTOGHG,Gg CO2 / yr,300,200,400
-6,USA,IPC2,TOTAL,CH4,Gg CH4 / yr,70,80,90
-7,USA,IPC3,TOTAL,CO2,Gg CO2 / yr,120,130,140
-8,FRA,IPC0,TOTAL,KYOTOGHG,Gg CO2 / yr,30,20,40
-9,FRA,IPC2,TOTAL,CH4,Gg CH4 / yr,7,8,9
-10,FRA,IPC2,TOTAL,CO2,Gg CO2 / yr,12,13,14
